pyproject.toml
[tool.poetry]
name = "ponzu-bot"
version = "0.1.0"
description = ""
authors = ["KiyotakaMatsushita <kiyotaka.86@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.25.1"
boto3 = "^1.34.98"
slack-bolt = "^1.16.4"
Flask = "^2.2.3"
python-dotenv = "^1.0.0"
glom = "^23.1.1"
langchain = "^0.1.17"
langgraph = "^0.0.44"
llama-index = "^0.10.34"
langchain-openai = "^0.1.6"
langchainhub = "^0.1.15"
urllib3 = "1.25.4"
watchdog = "^5.0.2"
google-auth = "^2.35.0"
google-cloud-firestore = "^2.18.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
serve = "ponzu_bot.app:main"
